--- src/util/InfiniteScroll.ts
@@ -1,18 +1,0 @@
-export const infiniteScroll = (target: string, fn: CallableFunction) => {
-  const options = {
-    threshold: 0.3,
-  };
-  const io = new IntersectionObserver((entries) => {
-    entries.forEach((entry) => {
-      if (!entry.isIntersecting) return;
-      fn().then(() => {
-        const eventTarget = document.querySelector(target);
-        io.disconnect();
-        if (eventTarget) io.observe(eventTarget);
-      });
-    });
-  }, options);
-  const last = document.querySelector(target);
-
-  if (last) io.observe(last);
-};
